additional functions, us_39144 split
