+++ config/splat.us.bobo6.yaml
@@ -69,6 +69,7 @@ segments:
       - [0x264BC, .rodata, e_stage_name]
       - [0x264F0, .rodata, richter]
       - [0x26C18, .rodata, us_39144]
+      - [0x26E20, .rodata, us_3E79C]
       - [0x27044, .rodata, prim_helpers]
       - [0x2704C, .rodata, e_life_up]
       - [0x27068, c, e_room_bg]
@@ -90,6 +91,7 @@ segments:
       - [0x34BD0, c, richter]
       - [0x39134, c, unused]
       - [0x39144, c, us_39144]
+      - [0x3E79C, c, us_3E79C]
       - [0x4D674, c, prim_helpers]
       - [0x4DEA4, c, e_life_up]
       - [0x4E758, bss] # unused
