+++ config/symbols.us.bobo6.txt
@@ -61,42 +61,6 @@ g_RelicOrbTextBg2EY = 0x80181010;
 g_RelicOrbSparkleX = 0x80181020;
 g_RelicOrbSparkleY = 0x80181030;
 BO6_rooms = 0x80181204;
-RIC_poseTimer = 0x8007632A;
-RIC_animSet = 0x8007632C;
-BO6_pStObjLayoutHorizontal = 0x80180180;
-BO6_pStObjLayoutVertical = 0x80180254;
-BO6_EntityUpdates = 0x80180328;
-BO6_EInitBreakable = 0x80180394;
-g_EInitObtainable = 0x801803A0;
-g_EInitParticle = 0x801803AC;
-g_EInitSpawner = 0x801803B8;
-g_EInitInteractable = 0x801803C4;
-g_EInitUnkId13 = 0x801803D0;
-g_EInitLockCamera = 0x801803DC;
-g_EInitCommon = 0x801803E8;
-g_EInitDamageNum = 0x0801803F4;
-BO6_BackgroundBlockInit = 0x801804E8;
-BO6_RedDoorTiles = 0x80180510;
-BO6_PrizeDrops = 0x80180610;
-D_80180DBC = 0x80180E1C;
-D_80180DD0 = 0x80180E28;
-aluric_subweapons_idx = 0x80180E38;
-aluric_subweapons_id = 0x80180E4C;
-g_goldCollectTexts = 0x80180E60;
-c_GoldPrizes = 0x80180E88;
-g_SubweaponAnimPrizeDrop = 0x80180EB0;
-D_80180EB8 = 0x80180F10;
-c_HeartPrizes = 0x80180F18;
-g_ExplosionYVelocities = 0x80180F1C;
-g_bigRedFireballAnim = 0x80180F44;
-g_ExplosionAnimations = 0x80180FC8;
-g_RelicOrbTexts = 0x80180FDC;
-g_RelicOrbTextBg1EY = 0x80180FE0;
-g_RelicOrbTextBg1SY = 0x80180FF0;
-g_RelicOrbTextBg2SY = 0x80181000;
-g_RelicOrbTextBg2EY = 0x80181010;
-g_RelicOrbSparkleX = 0x80181020;
-g_RelicOrbSparkleY = 0x80181030;
 subweapons_def = 0x801812CC;
 ric_anim_stand = 0x80181F44;
 BO6_EntityBackgroundBlock = 0x801A7068;
@@ -158,7 +122,6 @@ BO6_RicStepHighJump = 0x801B8F78;
 BO6_RicSetStep = 0x801B9928;
 BO6_RicSetAnimation = 0x801B9940;
 DecelerateX = 0x801B995C;
-DecelerateY = 0x801B99A4;
 BO6_RicCheckFacing = 0x801B99EC;
 BO6_RicSetSpeedX = 0x801B9AA4;
 BO6_RicSetInvincibilityFrames = 0x801B9AF4;
@@ -252,52 +215,3 @@ g_Ric_unk6C = 0x801D1684;
 g_Ric_unk70 = 0x801D1688;
 g_IsCutsceneDone = 0x801D1698;
 g_ItemIconSlots = 0x801D16A8;
-NO1Unused_801B7D24 = 0x801B9134;
-WarpBackgroundBrightness = 0x801C32BC;
-UnkPrimHelper = 0x801CD674;
-UpdateAnimation = 0x801CDA5C;
-FindFirstUnkPrim = 0x801CDBAC;
-FindFirstUnkPrim2 = 0x801CDBDC;
-PrimToggleVisibility = 0x801CDC58;
-PrimResetNext = 0x801CDCE0;
-UnkPolyFunc2 = 0x801CDDC0;
-UnkPolyFunc0 = 0x801CDE14;
-PrimDecreaseBrightness = 0x801CDE40;
-EntityLifeUpSpawn = 0x801CDEA4;
-g_Ric = 0x801D12C8; // size=0x3D0
-g_Ric_padPressed = 0x801D15E0;
-g_Ric_padTapped = 0x801D15E4;
-g_Ric_padHeld = 0x801D15E8;
-g_Ric_padSim = 0x801D15EC;
-g_Ric_demo_timer = 0x801D15F4;
-g_Ric_timers_0 = 0x801D15F8;
-g_Ric_timers_4 = 0x801D1600;
-g_Ric_timers_9 = 0x801D160A;
-g_Ric_timers_10 = 0x801D160C;
-g_Ric_timers_12 = 0x801D1610;
-g_Ric_timers_13 = 0x801D1612;
-g_Ric_timers_14 = 0x801D1614;
-g_Ric_vram_flag = 0x801D1618;
-g_Ric_unk04 = 0x801D161C;
-g_Ric_unk08 = 0x801D1620;
-g_Ric_status = 0x801D1624;
-g_Ric_unk40 = 0x801D1658;
-g_Ric_unk44 = 0x801D165C;
-g_Ric_unk4C = 0x801D1664;
-g_Ric_prev_step = 0x801D1668;
-g_Ric_prev_step_s = 0x801D166A;
-g_Ric_unk5C = 0x801D1674;
-g_Ric_unk6A = 0x801D1682;
-g_Ric_unk6C = 0x801D1684;
-g_Ric_unk70 = 0x801D1688;
-g_ItemIconSlots = 0x801D16A8;
-EntityShaft = 0x801C0D90;
-BO6_RicEntityVibhutiCrashCloud = 0x801C7FEC;
-BO6_rooms = 0x80181204;
-RIC_unk5A = 0x80076332;
-ric_anim_stand = 0x80181F44;
-DecelerateX = 0x801B995C;
-DecelerateY = 0x801B99A4;
-BO6_RicSetDebug = 0x801B9C3C;
-BO6_RicSetDeadPrologue = 0x801BA880;
-BO6_RicDoCrash = 0x801BA598;
